+++ src/Equinox.EventStore/EventStore.fs
@@ -451,8 +451,8 @@ module Caching =
             let! tokenAndState = load
             return! intercept streamName tokenAndState }
         interface ICategory<'event, 'state, string> with
-            member __.Load(log, streamName : string) : Async<StreamToken * 'state> =
-                loadAndIntercept (inner.Load(log, streamName)) streamName
+            member __.Load(log, streamName : string, opt) : Async<StreamToken * 'state> =
+                loadAndIntercept (inner.Load(log, streamName, opt)) streamName
             member __.TrySync(log : ILogger, (Token.StreamPos (stream,_) as token), state, events : 'event list) : Async<SyncResult<'state>> = async {
                 let! syncRes = inner.TrySync(log, token, state, events)
                 match syncRes with
@@ -474,12 +474,13 @@ module Caching =
 type private Folder<'event, 'state>(category : Category<'event, 'state>, fold: 'state -> 'event seq -> 'state, initial: 'state, ?readCache) =
     let batched log streamName = category.Load fold initial streamName log
     interface ICategory<'event, 'state, string> with
-        member __.Load(log, streamName) : Async<StreamToken * 'state> =
+        member __.Load(log, streamName, opt) : Async<StreamToken * 'state> =
             match readCache with
             | None -> batched log streamName
             | Some (cache : ICache, prefix : string) -> async {
                 match! cache.TryGet(prefix + streamName) with
                 | None -> return! batched log streamName
+                | Some tokenAndState when opt = Some AllowStale -> return tokenAndState
                 | Some (token, state) -> return! category.LoadFromToken fold state streamName token log }
         member __.TrySync(log : ILogger, token, initialState, events : 'event list) : Async<SyncResult<'state>> = async {
             let! syncRes = category.TrySync fold log (token, initialState) events
@@ -524,12 +525,12 @@ type Resolver<'event,'state>
     let resolveTarget = function AggregateId (cat,streamId) -> sprintf "%s-%s" cat streamId | StreamName streamName -> streamName
     member __.Resolve(target, [<O; D null>]?option) =
         match resolveTarget target, option with
-        | sn,None -> resolveStream sn
-        | sn,Some AssumeEmpty -> Stream.ofMemento (context.LoadEmpty sn,initial) (resolveStream sn)
+        | sn,(None|Some AllowStale) -> resolveStream option sn
+        | sn,Some AssumeEmpty -> Stream.ofMemento (context.LoadEmpty sn,initial) (resolveStream option sn)
 
     /// Resolve from a Memento being used in a Continuation [based on position and state typically from Stream.CreateMemento]
     member __.FromMemento(Token.Unpack token as streamToken, state) =
-        Stream.ofMemento (streamToken,state) (resolveStream token.stream.name)
+        Stream.ofMemento (streamToken,state) (resolveStream None token.stream.name)
 
 type private SerilogAdapter(log : ILogger) =
     interface EventStore.ClientAPI.ILogger with